working float print, new USB serial printing lib
